+++ src/time.rs
@@ -1,4 +1,31 @@
 //! Time units
+//!
+//! See [`Hertz`], [`KiloHertz`] and [`MegaHertz`] for creating increasingly higher frequencies.
+//!
+//! The [`U32Ext`] trait adds various methods like `.hz()`, `.mhz()`, etc to the `u32` primitive type,
+//! allowing it to be converted into frequencies.
+//!
+//! # Examples
+//!
+//! ## Create a 2 MHz frequency
+//!
+//! This example demonstrates various ways of creating a 2 MHz (2_000_000 Hz) frequency. They are
+//! all equivalent, however the `2.mhz()` variant should be preferred for readability.
+//!
+//! ```rust
+//! use stm32f1xx_hal::{
+//!     time::Hertz,
+//!     // Imports U32Ext trait
+//!     prelude::*,
+//! };
+//!
+//! let freq_hz = 2_000_000.hz();
+//! let freq_khz = 2_000.khz();
+//! let freq_mhz = 2.mhz();
+//!
+//! assert_eq!(freq_hz, freq_khz);
+//! assert_eq!(freq_khz, freq_mhz);
+//! ```
 
 use cortex_m::peripheral::DWT;
 
@@ -9,20 +36,73 @@ use crate::rcc::Clocks;
 pub struct Bps(pub u32);
 
 /// Hertz
+///
+/// Create a frequency specified in [Hertz](https://en.wikipedia.org/wiki/Hertz).
+///
+/// See also [`KiloHertz`] and [`MegaHertz`] for semantically correct ways of creating higher
+/// frequencies.
+///
+/// # Examples
+///
+/// ## Create an 60 Hz frequency
+///
+/// ```rust
+/// use stm32f1xx_hal::time::Hertz;
+///
+/// let freq = 60.hz();
+/// ```
 #[derive(Clone, Copy, PartialEq, Debug)]
 pub struct Hertz(pub u32);
 
-/// KiloHertz
+/// Kilohertz
+///
+/// Create a frequency specified in kilohertz.
+///
+/// See also [`Hertz`] and [`MegaHertz`] for semantically correct ways of creating lower or higher
+/// frequencies.
+///
+/// # Examples
+///
+/// ## Create a 100 Khz frequency
+///
+/// This example creates a 100 KHz frequency. This could be used to set an I2C data rate or PWM
+/// frequency, etc.
+///
+/// ```rust
+/// use stm32f1xx_hal::time::Hertz;
+///
+/// let freq = 100.khz();
+/// ```
 #[derive(Clone, Copy, PartialEq, Debug)]
 pub struct KiloHertz(pub u32);
 
-/// MegaHertz
+/// Megahertz
+///
+/// Create a frequency specified in kilohertz.
+///
+/// See also [`Hertz`] and [`KiloHertz`] for semantically correct ways of creating lower
+/// frequencies.
+///
+/// # Examples
+///
+/// ## Create a an 8 MHz frequency
+///
+/// This example creates an 8 MHz frequency that could be used to configure an SPI peripheral, etc.
+///
+/// ```rust
+/// use stm32f1xx_hal::time::Hertz;
+///
+/// let freq = 8.mhz();
+/// ```
 #[derive(Clone, Copy, PartialEq, Debug)]
 pub struct MegaHertz(pub u32);
 
 /// Time unit
 #[derive(PartialEq, PartialOrd, Clone, Copy)]
 pub struct MilliSeconds(pub u32);
+
+#[derive(PartialEq, PartialOrd, Clone, Copy)]
+pub struct MicroSeconds(pub u32);
 
 /// Extension trait that adds convenience methods to the `u32` type
 pub trait U32Ext {
@@ -40,6 +120,9 @@ pub trait U32Ext {
 
     /// Wrap in `MilliSeconds`
     fn ms(self) -> MilliSeconds;
+
+    /// Wrap in `MicroSeconds`
+    fn us(self) -> MicroSeconds;
 }
 
 impl U32Ext for u32 {
@@ -62,23 +145,39 @@ impl U32Ext for u32 {
     fn ms(self) -> MilliSeconds {
         MilliSeconds(self)
     }
-}
 
-impl Into<Hertz> for KiloHertz {
-    fn into(self) -> Hertz {
-        Hertz(self.0 * 1_000)
+    fn us(self) -> MicroSeconds {
+        MicroSeconds(self)
     }
 }
 
-impl Into<Hertz> for MegaHertz {
-    fn into(self) -> Hertz {
-        Hertz(self.0 * 1_000_000)
+impl From<KiloHertz> for Hertz {
+    fn from(val: KiloHertz) -> Self {
+        Self(val.0 * 1_000)
     }
 }
 
-impl Into<KiloHertz> for MegaHertz {
-    fn into(self) -> KiloHertz {
-        KiloHertz(self.0 * 1_000)
+impl From<MegaHertz> for Hertz {
+    fn from(val: MegaHertz) -> Self {
+        Self(val.0 * 1_000_000)
+    }
+}
+
+impl From<MegaHertz> for KiloHertz {
+    fn from(val: MegaHertz) -> Self {
+        Self(val.0 * 1_000)
+    }
+}
+
+impl Into<Hertz> for MilliSeconds {
+    fn into(self) -> Hertz {
+        Hertz(1_000 / self.0)
+    }
+}
+
+impl Into<Hertz> for MicroSeconds {
+    fn into(self) -> Hertz {
+        Hertz(1_000_000 / self.0)
     }
 }